update: 인증이 필요없는 게시글 api 세분화
