+++ src/main/java/com/fithub/fithubbackend/domain/board/api/PostController.java
@@ -1,7 +1,10 @@
 package com.fithub.fithubbackend.domain.board.api;
 
+import com.fithub.fithubbackend.domain.board.application.CommentService;
 import com.fithub.fithubbackend.domain.board.application.PostService;
 import com.fithub.fithubbackend.domain.board.dto.*;
+import com.fithub.fithubbackend.domain.board.dto.comment.CommentInfoDto;
+import com.fithub.fithubbackend.domain.board.dto.comment.ParentCommentInfoDto;
 import com.fithub.fithubbackend.domain.board.dto.likes.LikedUsersInfoDto;
 import com.fithub.fithubbackend.global.exception.ErrorResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
@@ -27,6 +30,7 @@ import java.util.List;
 public class PostController {
 
     private final PostService postService;
+    private final CommentService commentService;
 
     @Operation(summary = "게시글 전체 조회, page 사용 (size = 9, sort = \"id\", desc 적용). 페이지 이동 시 page 값만 보내주면 됨. ex) \"page\" : 0 인 경우 1 페이지", responses = {
             @ApiResponse(responseCode = "200", description = "게시글 전체 조회 성공"),
@@ -65,6 +69,28 @@ public class PostController {
     @PostMapping("/likes")
     public ResponseEntity<List<LikedUsersInfoDto>> getLikedUsersForPostDetail(@RequestBody List<PostRequestDto> postRequestDtos) {
         return ResponseEntity.ok(postService.getLikedUsersForPosts(postRequestDtos));
+    }
+
+    @Operation(summary = "최상위 댓글 조회. page 사용. (size = 10, sort = \"id\", desc 적용). 페이지 이동 시 page 값만 보내주면 됨. ex) \"page\" : 0 인 경우 1 페이지", responses = {
+            @ApiResponse(responseCode = "200", description = "최상위 댓글 조회 완료"),
+    }, parameters = {
+            @Parameter(name="postId", description = "게시글 id")
+    })
+    @GetMapping("/comments/{postId}")
+    public ResponseEntity<Page<ParentCommentInfoDto>> getComments(@PageableDefault(sort = "id", direction = Sort.Direction.DESC) Pageable pageable,
+                                                                  @PathVariable(value = "postId") long postId) {
+        return ResponseEntity.ok().body(commentService.getCommentsWithPage(pageable, postId));
+    }
+
+    @Operation(summary = "대댓글 조회", responses = {
+            @ApiResponse(responseCode = "200", description = "대댓글 조회 완료"),
+    }, parameters = {
+            @Parameter(name="postId", description = "게시글 id"),
+            @Parameter(name="commentId", description = "최상위 댓글 id")
+    })
+    @GetMapping("/comments/{postId}/{commentId}")
+    public ResponseEntity<List<CommentInfoDto>> getDetailComments(@PathVariable(value = "postId") long postId, @PathVariable(value = "commentId") long commentId) {
+        return ResponseEntity.ok().body(commentService.getDetailComments(commentId));
     }
 
 }
